+++ testutil/Fixture/template/content-in-view.html.php
@@ -1,0 +1,7 @@
+<?php
+
+use SubstancePHP\HTTP\Renderer\HtmlRenderer;
+
+/** @var HtmlRenderer $this */
+?>
+[<?= $this->content() ?>]