+++ testutil/Fixture/template/partials/layout-meddler.html.php
@@ -1,0 +1,9 @@
+<?php
+
+use SubstancePHP\HTTP\Renderer\HtmlRenderer;
+
+/** @var HtmlRenderer $this */
+?>
+
+<?php $this->layout('other'); ?>
+<p>meddled</p>